+++ examples/ui-demo/src/hooks/7702/constants.ts
@@ -1,0 +1,3 @@
+export const ODYSSEY_EXPLORER_URL = "https://odyssey-explorer.ithaca.xyz";
+
+export const SESSION_KEY_VALIDITY_TIME_SECONDS: number = 6000; // 10 minutes
