+++ examples/ui-demo/src/hooks/7702/genEntityId.ts
@@ -1,0 +1,8 @@
+import { maxUint32 } from "viem";
+
+export function genEntityId(): number {
+  const min: number = 1;
+  const max: number = Number(maxUint32);
+
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+}
